# src/main/java/io/swagger/api/AlarmApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.27).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.Alarm;
import io.swagger.model.AlarmBody;
import io.swagger.model.AlarmCreatedResponse;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.CookieValue;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2021-08-12T06:02:13.164Z[GMT]")
@Validated
public interface AlarmApi {

    @Operation(summary = "", description = "", tags={  })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "Alarm activated", content = @Content(mediaType = "application/json", schema = @Schema(implementation = AlarmCreatedResponse.class))) })
    @RequestMapping(value = "/alarm/{alarmName}",
        produces = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<AlarmCreatedResponse> alarmAlarmNamePost(@Parameter(in = ParameterIn.PATH, description = "", required=true, schema=@Schema()) @PathVariable("alarmName") String alarmName);


    @Operation(summary = "", description = "", tags={ "alarms" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "Alarm activated", content = @Content(mediaType = "application/json", schema = @Schema(implementation = AlarmCreatedResponse.class))) })
    @RequestMapping(value = "/alarm",
        produces = { "application/json" }, 
        consumes = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<AlarmCreatedResponse> createAlarm(@Parameter(in = ParameterIn.DEFAULT, description = "alarm's data", required=true, schema=@Schema()) @Valid @RequestBody AlarmBody body);


    @Operation(summary = "", description = "", tags={ "alarms" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "Return all the created alarms", content = @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Alarm.class)))),
        
        @ApiResponse(responseCode = "400", description = "Invalid status value") })
    @RequestMapping(value = "/alarm",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<List<Alarm>> findAllAlarms();

}